SpreadsheetFormatterSelectorEditHateosHttpHandlerCell & SpreadsheetFormatterSelectorEditHateosHttpHandlerMetadata
- SpreadsheetFormatterSelectorEditHateosHttpHandlerCell: Context.cell will contain the SpreadsheetCell in the URL.
